Updated routes and connection

+++ controllers/api/comment-routes.js
@@ -11,6 +11,7 @@ router.get('/', (req, res) => {
 });
 
 router.post('/', (req, res) => {
+  if (req.session) {
     Comment.create({
         comment_text: req.body.comment_text,
         user_id: req.body.user_id,
@@ -21,10 +22,11 @@ router.post('/', (req, res) => {
           console.log(err);
           res.status(400).json(err);
         });
-
+  }
 });
 
 router.delete('/:id', (req, res) => {
+  if (req.session) {
     Comment.destroy({
         where: {
           id: req.params.id
@@ -41,6 +43,7 @@ router.delete('/:id', (req, res) => {
           console.log(err);
           res.status(500).json(err);
         });
+      }
 });
 
 module.exports = router;
